Sabre-Cli/Sabre-Web/Sabre-Cli/old/requirements.txt
pip:
cmd2
notify2
pexpect
#pyqterm


apt:
redis-server
python-redis
pip
#python-qt4

Simple implant Needs to be in path


TPT: NEEDS executables in path
metasploit
empire
